+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This means that:
  * Support of inheriting attributes of referred in objects/enums using `$ref`
  * Added support for array properties, `default: [...]` values is recognized and applied (but only for primitive types, objects not)
  * Introduced validation which makes sure only supported `x-keytiles-` vendor attributes are present in the schema - otherwise build will fail. This helps in future releases to detect accidental leftovers on schema files of `x-keytiles-` attributes which were deprecated and removed in the meantime can not be overlooked
+ * Added support for composition of Enums - see README for more details
 
 ## Bugfixes
  * During inheritance when Superclass has a constructor sometimes the Subclass did not pick up imports needed by Superclass - that lead to generated model compilation problems. This is now fixed (or at least better). Also added model structure to tested OpenApi schemas which is checking this works.